pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0.0",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "geofabrics"
version = "1.0.1"
description = "A package for creating geofabrics for flood modelling."
readme = "README.md"
authors = [{ name = "Rose pearson", email = "rose.pearson@niwa.co.nz" }]
license = { file = "LICENSE" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: GIS",
    "Programming Language :: Python :: 3 :: Only",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
]
keywords = ["GeoFabrics", "DEM", "Roughness", "Hydrologically conditioned", "LiDAR"]
dependencies = [
    "dask",
    "distributed",
    "gdal",
    "geopandas>=0.10",
    "pdal",
    "rioxarray",
    "geoapis",
    "netcdf4",
    "osmpythontools>=0.3.5",
    'tomli; python_version < "3.11"',
]
requires-python = ">=3.9"

[project.scripts]
geofabrics_from_file = "geofabrics.__main__:cli_run_from_file"
geofabrics_from_dict = "geofabrics.__main__:cli_run_from_dict"

[tool.setuptools]
packages = ["geofabrics"]

[project.optional-dependencies]
dev = ["black", "check-manifest", "python-dotenv", "pip-tools", "pytest"]

[project.urls]
Homepage = "https://github.com/rosepearson/GeoFabrics"
Documentation = "https://rosepearson.github.io/GeoFabrics/"
